Added json file (xkcd.path.json) to managed file paths and made necessary updates. This file will always be expected to be present in the current directory (./).
Updated xkcd.py to reflect the changes relative to the path.

Now use an entry in the config file to retrieve the latest comic
number, for dumb system like ubuntu which fail to retrieve it
correctly. I'm yet to discover the root of this bug.

Removed unrelated files like janitor.py to prepare the repository
to become exclusive to the xkcd bot.